Оправено да не изтрива като бъгне, добавен бутон Clear Settings на Right Click  върху компонент, за да може да изчисти ако има някакви добавени неща към него, които не се виждат :D, оправено Remove connections. ПОИГРАЙТЕ СИ!  :Р

+++ OOD2-master-2/OOD2_project/OOD2_project/Connection.cs
@@ -13,8 +13,8 @@ namespace OOD2_project
         public Component startComponent;
         public Component endComponent;
         public Color color;
-        public int flow;
-        private int maxFlow;
+        public double flow;
+        private double maxFlow;
         public Point[] curvePoints;
 
         public Connection(Component StartComp, Component EndComp, int MaxFlow, Point[] points)
